Refactor SearchBuilder search logic customizations for Blacklight 9.0.0.beta8
This seems to be what it requires of us. Under discussion in https://github.com/projectblacklight/blacklight/pull/3762 it is somewhat possible it will change before 9.0.0 final, although some people strongly prefer it this way.
