+++ app/controllers/featured_topic_controller.rb
@@ -4,7 +4,9 @@ class FeaturedTopicController < CatalogController
 
   configure_blacklight do |config|
     # Limit just to items in the featured topic.
+    # For Blacklight 9, we need to do this in two classes.
     config.search_builder_class = ::SearchBuilder::WithinFeaturedTopicBuilder
+    config.facet_search_builder_class = ::SearchBuilder::WithinFeaturedTopicFacetBuilder
 
     # And we need to make sure the :slug param is allowed by blacklight,
     # don't totally understand why, as of BL 7.25
