+++ app/models/facet_search_builder.rb
@@ -1,0 +1,7 @@
+class FacetSearchBuilder < Blacklight::FacetSearchBuilder
+  include Blacklight::Solr::FacetSearchBuilderBehavior
+
+  # shared logic with SearchBuilder that MUST be duplciated to avoid
+  # subtle faulty behavior, see https://github.com/projectblacklight/blacklight/pull/3762
+  include SearchBuilderBehavior
+end
