+++ app/models/search_builder/within_featured_topic_builder.rb
@@ -4,15 +4,8 @@ class SearchBuilder
   # :slug needs to be provided in context.
   #
   # Used on FeaturedTopicController.
+  #
   class WithinFeaturedTopicBuilder < ::SearchBuilder
-    extend ActiveSupport::Concern
-
-    self.default_processor_chain += [:within_featured_topic]
-
-    def within_featured_topic(solr_parameters)
-      featured_topic = FeaturedTopic.from_slug(blacklight_params[:slug])
-      solr_parameters[:fq] ||= []
-      solr_parameters[:fq] << featured_topic.solr_fq
-    end
+    include WithinFeaturedTopicBuilderBehavior
   end
 end
